pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "psyflow"
version = "0.1.3"
description = "A utility package for building modular PsychoPy experiments."
requires-python = ">=3.10"
authors = [
    { name="Zhipeng Cao", email="zhipeng30@foxmail.com" }
]
dependencies = [
    "psychopy",
    "numpy",
    "pandas",
    "click",
    "cookiecutter",
    "pyyaml",
    "pyserial",
    "edge-tts",
]
dynamic = ["entry-points"]

[project.scripts]
psyflow = "psyflow.cli:main"
